wrestling with an aligator
Attempting to get this diary saving shit worked out.
Turns out we need to use basic-ass file load and no Marshal stuff because that works with scrambled files instead of plaintext files, and I prefer to work in plaintext because honestly, it's just a fucking list of picture names, there's no need to keep that under heavy wraps.
Question is how do we retrieve save file number, how do we shove the loaded data into the array properly, and eventually, how do we save the data in a plaintext file in a manner similar to the example load file?

-turned down the volume on those fucking ding noises in the healing animation. ow.

As a reminder, friday is the day to begin Serious Work on this heavy stuff. While I'll prolly be working on and off between then (since thursday is usually my best free day instead of friday due to work schedule), that is when Nyx mentioned she would be free to work at full-cylinders on this.

-special thanks to psy_wombats and delsin7 for helping my clueless ass through this file shit-