taking a walk
-minor details added to dinner scene
-started work on event movement for walk back to crossroads after dinner
scene
-added a chimney
-added slight details to artichoke event
-moved a tall tree in the lakeside mid-area
-stuck a save point next to the house because im tired of walking all
the way there to test things
-for some reason the github desktop program I usually use shat itself
and died so now I have to use the back-up of the old one to send this.
>_>

to-do:
-finish the walking directions
-write some actual lines for the walking section
-make francis waving sprite
-make francis fond glance sprite
-start planning the ever-so-slight differences present in the lunch
version

(note: I really need to get my ass moving on finishing jello and the
valley of duh. I wish I didn't feel double obligation guilt for
abandoning one for the other back and forth no matter which project I
choose to work on.....)